Put all dtype conversion stuff into the _misc namespaces
